PTS should be passed as int64_t to demux_ty_CopyToDemuxPacket
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@24517 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/demux_ty.c
@@ -334,7 +334,7 @@ static void demux_ty_AddToAudioBuffer( TiVoInfo *tivo, unsigned char *buffer,
 }
 
 static void demux_ty_CopyToDemuxPacket( demux_stream_t *ds,
-       unsigned char *buffer, int size, off_t pos, float pts )
+       unsigned char *buffer, int size, off_t pos, int64_t pts )
 {
    demux_packet_t *dp = new_demux_packet( size );
    memcpy( dp->buffer, buffer, size );
